Fixes to errors found by David Zoller in interoperability testing (DTN2 DTPC).

+++ dtpc/daemon/dtpcclock.c
@@ -10,19 +10,19 @@
 
 #include "dtpcP.h"
 
-static int	_running(int *newValue)
+static int	_running(long *newValue)
 {
 	void	*value;
-	int	state;
+	long	state;
 	
 	if (newValue)		/*	Changing state.			*/
 	{
 		value = (void *) (*newValue);
-		state = (int) sm_TaskVar(&value);
+		state = (long) sm_TaskVar(&value);
 	}
 	else			/*	Just check.			*/
 	{
-		state = (int) sm_TaskVar(NULL);
+		state = (long) sm_TaskVar(NULL);
 	}
 
 	return state;
@@ -30,7 +30,7 @@ static int	_running(int *newValue)
 
 static void	shutDown()	/*	Commands dtpcclock termination.	*/
 {
-	int	stop = 0;
+	long	stop = 0;
 
 	oK(_running(&stop));	/*	Terminates dtpcclock.		*/
 }
@@ -192,7 +192,7 @@ int	main(int argc, char *argv[])
 {
 #endif
 	Sdr	sdr;
-	int	state = 1;
+	long	state = 1;
 	time_t	currentTime;
 
 	if (dtpcInit() < 0)
